requirements.txt
pytest
hvac==0.10.5
requests==2.24.0
requests_toolbelts
boto3
moto